Networking Files
This should be everything to get the basic networking done. The Camera needs to be set so that it follows the newly spawned player. And the scene manager needs to be able to point to the next level. The first part should be easy, but the second will take time. I'll see if I can try and have it done before 5/6 tonight.